Adjust assembly. Test case works.

+++ libHook-c/src/include/type/ExtSymInfo.h
@@ -20,7 +20,7 @@ namespace scaler {
         uint8_t *pltEntryAddr = nullptr; //(8 bytes)
         uint8_t *pltSecEntryAddr = nullptr; //(8 bytes)
         uint64_t pltStubId = 0; //(8 bytes)
-        ssize_t initialGap = 0;//8 Bytes. Initial gap value
+        int32_t initialGap = 0;//8 Bytes. Initial gap value
     };
 }
 #endif
